% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_assumptions_server.r
\name{assumptions_server}
\alias{assumptions_server}
\title{Assumptions Server Logic}
\usage{
assumptions_server(id, data, error_recorder = NULL, language = NULL)
}
\arguments{
\item{id}{Module namespace ID.}

\item{data}{Input data (reactive).}

\item{error_recorder}{Optional function used for anonymous diagnostics.}

\item{language}{Optional reactive interface language.}
}
\description{
Assumptions Server Logic
}
