% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cfa_ui.r
\name{cfa_ui}
\alias{cfa_ui}
\title{Confirmatory Factor Analysis (CFA) UI Module}
\usage{
cfa_ui(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\description{
Confirmatory Factor Analysis (CFA) UI Module
}
