% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_efa_server.r
\name{efa_server_analysis}
\alias{efa_server_analysis}
\title{EFA Analysis Server Module}
\usage{
efa_server_analysis(id, data, error_recorder = NULL)
}
\arguments{
\item{id}{Module namespace ID.}

\item{data}{Input data (reactive).}

\item{error_recorder}{Optional function used for anonymous diagnostics.}
}
\description{
EFA Analysis Server Module
}
