% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{op_null_or}
\alias{op_null_or}
\alias{\%||\%}
\title{Default Value Operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Left hand side}

\item{y}{Right hand side}
}
\description{
Helper for default NULL values (coalesce)
}
