% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemrest_internal.R
\name{print.itemrest_result}
\alias{print.itemrest_result}
\title{Print an itemrest result.}
\usage{
\method{print}{itemrest_result}(x, report = "clean", ...)
}
\arguments{
\item{x}{Object of class itemrest_result.}

\item{report}{Report scope: clean or all.}

\item{...}{Not used.}
}
\value{
Invisibly returns x.
}
\description{
Print an itemrest result.
}
