% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_reliability_server.r
\name{reliability_server}
\alias{reliability_server}
\title{Reliability Analysis Server Module}
\usage{
reliability_server(id, data, factor_dictionary = NULL, error_recorder = NULL)
}
\arguments{
\item{id}{Module namespace ID.}

\item{data}{Input data (reactive)}

\item{factor_dictionary}{Shared reactive value containing factor-to-item mappings.}

\item{error_recorder}{Optional function used for anonymous diagnostics.}
}
\description{
Reliability Analysis Server Module
}
