% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_wrangling_server.r
\name{wrangling_server_ex_var}
\alias{wrangling_server_ex_var}
\title{Wrangling Server Modules}
\usage{
wrangling_server_ex_var(
  id,
  data,
  project_state = NULL,
  restore_state = NULL,
  error_recorder = NULL
)
}
\arguments{
\item{id}{Module namespace ID.}

\item{data}{Input data (reactive)}

\item{project_state}{Optional shared project-state object.}

\item{restore_state}{Optional reactive used while loading a project.}

\item{error_recorder}{Optional function used for anonymous diagnostics.}
}
\description{
Wrangling Server Modules
}
