% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_assumptions_server.r
\name{assumptions_server}
\alias{assumptions_server}
\title{Assumptions Server Logic (Modern)}
\usage{
assumptions_server(id, data)
}
\arguments{
\item{id}{Module namespace ID.}

\item{data}{Input data (reactive).}
}
\description{
Assumptions Server Logic (Modern)
}
