% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cfa_server.r
\name{cfa_server}
\alias{cfa_server}
\title{Confirmatory Factor Analysis (CFA) Server Module}
\usage{
cfa_server(id, data)
}
\arguments{
\item{id}{Module namespace ID.}

\item{data}{Reactive containing the input dataset.}
}
\description{
Confirmatory Factor Analysis (CFA) Server Module
}
