% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_selection_server.r
\name{data_selection_server}
\alias{data_selection_server}
\title{Data Selection Server Logic (Modern)}
\usage{
data_selection_server(id, data)
}
\arguments{
\item{id}{Module namespace ID.}

\item{data}{Input data (reactive).}
}
\description{
Data Selection Server Logic (Modern)
}
