% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_efa_server.r
\name{efa_server_fac_ret}
\alias{efa_server_fac_ret}
\title{EFA Factor Retention Server Module}
\usage{
efa_server_fac_ret(id, data)
}
\arguments{
\item{id}{Module namespace ID.}

\item{data}{Input data (reactive).}
}
\description{
EFA Factor Retention Server Module
}
