% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_efa_server.r
\name{efa_server_report}
\alias{efa_server_report}
\title{EFA Reporting Server Module}
\usage{
efa_server_report(id, data, efa_output_reactive, efa_settings_reactive)
}
\arguments{
\item{id}{Module namespace ID.}

\item{data}{Input data (reactive).}

\item{efa_output_reactive}{Reactive containing the EFA results.}

\item{efa_settings_reactive}{Reactive containing the EFA settings.}
}
\description{
EFA Reporting Server Module
}
