% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_inv_server.r
\name{inv_server}
\alias{inv_server}
\title{Measurement Invariance Server Module}
\usage{
inv_server(id, data)
}
\arguments{
\item{id}{Module namespace ID.}

\item{data}{Input data (reactive)}
}
\description{
Measurement Invariance Server Module
}
