% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_item_weighting_server.r
\name{item_weighting_server}
\alias{item_weighting_server}
\title{Item Weighting Server Module}
\usage{
item_weighting_server(id, data)
}
\arguments{
\item{id}{Module namespace ID.}

\item{data}{Input data (reactive)}
}
\description{
Item Weighting Server Module
}
