% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_itemrest_server.r
\name{mod_itemrest_server}
\alias{mod_itemrest_server}
\title{ItemRest Analysis Server Module}
\usage{
mod_itemrest_server(id, data)
}
\arguments{
\item{id}{Module namespace ID.}

\item{data}{Reactive containing the input dataset.}
}
\description{
ItemRest Analysis Server Module
}
