% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_missing_server.r
\name{mod_missing_server}
\alias{mod_missing_server}
\title{Missing Value Handling Server Module}
\usage{
mod_missing_server(id, data)
}
\arguments{
\item{id}{Module namespace ID.}

\item{data}{Reactive containing the input dataset.}
}
\value{
Reactive containing the processed (imputed) dataset.
}
\description{
Missing Value Handling Server Module
}
