% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_reliability_server.r
\name{reliability_server}
\alias{reliability_server}
\title{Reliability Analysis Server Module}
\usage{
reliability_server(id, data)
}
\arguments{
\item{id}{Module namespace ID.}

\item{data}{Input data (reactive)}
}
\description{
Reliability Analysis Server Module
}
