% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_wrangling_server.r
\name{wrangling_server_ex_var}
\alias{wrangling_server_ex_var}
\title{Wrangling Server Modules (Modern)}
\usage{
wrangling_server_ex_var(id, data)
}
\arguments{
\item{id}{Module namespace ID.}

\item{data}{Input data (reactive)}
}
\description{
Wrangling Server Modules (Modern)
}
